test(VerifyCertificate): prevent JSONLD failure by stubbing verification process
There is a problem with the babel configuration for tests. It seems to take the code from CVJS and for some reason avoid expanding the contexts, which remain strings (URLs), thus failing normalization. The issue is triggered by the "loose" flag of the babel config. Unfortunately, removing it actually breaks other tests (ie: cannot set a stub to a "getter" (instead of function) from domain.certificates). So we reduce the integration test CVJS <> BUV by stubbing the whole verify function.